Refactor project to maven project
