--- README.md
@@ -1,6 +1,0 @@
-## 홍익대학교 윤영 교수님 소프트웨어 공학 수업 Unit Test 및 XP 짝 프로그래밍 실습
-
-## 팀원
-- B911059 노선우(Book Class 코드 작성)
-- B911021 김동영(Test Code 작성)
-- B911029 김민중(Test Code 작성)
